Update Test.java
Changed 'int' to 'long' on line 23 in case reversed number is > maximum value for an int.

+++ Test.java
@@ -20,7 +20,7 @@ public class Test {
 	}
 	
 	private static int reverseSecondApproach(int input) {
-		int newNum = 0;
+		long newNum = 0;
 		while(input > 0) {
 			newNum *= 10;
 			newNum += input % 10;
